@@ -1,7 +1,7 @@
 {
   "name": "nguniversal",
   "main": "index.js",
-  "version": "5.0.0-beta.5",
+  "version": "6.0.0",
   "description": "Universal (isomorphic) JavaScript support for Angular",
   "homepage": "https://github.com/angular/universal",
   "license": "MIT",
@@ -59,25 +59,38 @@
   "scripts": {
     "lint": "tslint -c tslint.json '+(modules|tools)/**/*.ts' --exclude '**/node_modules/**/*'",
     "build": "ts-node --project ./tools/tsconfig.json build.ts",
-    "test:watch": "ts-node --project ./tools/tsconfig.json build-test.ts && karma start karma.conf.js",
-    "test:browser": "ts-node --project ./tools/tsconfig.json build-test.ts && ts-node test/test-browser.ts",
-    "test:server": "ts-node --project ./tools/tsconfig.json build-test.ts && ts-node test/test-server.ts"
+    "upver": "standard-version",
+    "upver:beta": "standard-version --prerelease beta",
+    "upver:rc": "standard-version --prerelease rc",
+    "upver:push": "standard-version",
+    "postinstall": "ngc -p angular.tsconfig.json",
+    "prebuildifier": "bazel build --noshow_progress @com_github_bazelbuild_buildtools//buildifier",
+    "buildifier": "find . -type f \\( -name BUILD -or -name BUILD.bazel \\) ! -path \"*/node_modules/*\" | xargs $(bazel info bazel-bin)/external/com_github_bazelbuild_buildtools/buildifier/*/buildifier",
+    "bazel:build": "bazel build //...",
+    "test": "bazel test //...",
+    "build:watch": "ibazel build //...",
+    "test:watch": "ibazel test //..."
   },
   "devDependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/platform-server": "^5.2.0",
+    "@angular/animations": "^6.0.0",
+    "@angular/bazel": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/http": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@angular/platform-server": "^6.0.0",
+    "@bazel/ibazel": "^0.3.1",
     "@types/express": "^4.0.39",
+    "@types/fs-extra": "^4.0.5",
     "@types/hapi": "^16.1.2",
     "@types/jasmine": "^2.8.6",
+    "@types/node": "^9.4.6",
     "camelcase": "^4.1.0",
     "express": "^4.15.2",
+    "fs-extra": "^3.0.1",
     "glob": "^7.1.2",
     "hapi": "^16.1.1",
     "jasmine-core": "^2.8.0",
@@ -88,20 +101,24 @@
     "karma-typescript": "^3.0.12",
     "minimatch": "^3.0.4",
     "protractor": "^5.2.0",
+    "replace-in-file": "^3.1.1",
     "rimraf": "^2.6.1",
-    "rollup": "~0.41.6",
+    "rollup": "^0.56.5",
     "rollup-plugin-alias": "^1.4.0",
     "rollup-plugin-commonjs": "^8.2.6",
     "rollup-plugin-node-resolve": "^3.0.2",
     "rollup-plugin-sourcemaps": "^0.4.2",
     "rollup-plugin-uglify": "^2.0.1",
-    "rxjs": "^5.5.2",
+    "rxjs": "^6.0.0",
+    "sorcery": "^0.10.0",
+    "standard-version": "^4.3.0",
     "systemjs": "0.19.43",
     "ts-node": "^3.0.4",
     "tsconfig-paths": "^2.3.0",
     "tslint": "^5.9.1",
-    "tsutils": "^2.6.0",
-    "typescript": "~2.6.2",
+    "tsutils": "^2.21.2",
+    "typescript": "~2.7.2",
+    "uglify-js": "^2.8.14",
     "zone.js": "^0.8.12"
   }
 }
