@@ -3,10 +3,14 @@ import * as ts from 'typescript';
 import * as Lint from 'tslint';
 import * as minimatch from 'minimatch';
 
-const buildConfig = require('../../build-config');
-
 /** License banner that is placed at the top of every public TypeScript file. */
-const licenseBanner = buildConfig.licenseBanner;
+const licenseBanner = `/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */`;
 
 /** Failure message that will be shown if a license banner is missing. */
 const ERROR_MESSAGE = 'Missing license header in this TypeScript file. ' +
